llm_sentiment/sentiment/requirements.txt
pandas
requests
PyYAML
beautifulsoup4
lxml
transformers
torch
pyarrow
fastparquet
python-dateutil
tqdm
fastapi 
uvicorn
pydantic_settings
clickhouse-driver
vaderSentiment
apscheduler